+++ test/org/jetbrains/plugins/ideavim/action/ChangeActionTest.java
@@ -135,7 +135,17 @@ public class ChangeActionTest extends VimTestCase {
 
   // VIM-200 |c| |w|
   public void testChangeWordAtLastChar() {
-    doTest(parseKeys("cw"), "on<caret>e two three\n", "on two three\n");
+    doTest(parseKeys("cw"), "on<caret>e two three\n", "on<caret> two three\n");
+  }
+
+  // VIM-1380 |c| |w| |count|
+  public void testChangeTwoWordsAtLastChar() {
+    doTest(parseKeys("c2w"), "on<caret>e two three\n", "on<caret> three\n");
+  }
+
+  // VIM-1380 |d| |w| |count|
+  public void testDeleteTwoWordsAtLastChar() {
+    doTest(parseKeys("d2w"), "on<caret>e two three\n", "on<caret>three\n");
   }
 
   // VIM-515 |c| |W|
@@ -377,6 +387,50 @@ public class ChangeActionTest extends VimTestCase {
            "br\n");
   }
 
+  // VIM-1379 |CTRL-V| |j| |v_b_I|
+  public void testInsertVisualBlockWithEmptyLineInTheMiddle() {
+    doTest(parseKeys("ll", "<C-V>", "jjI", "_quux_", "<Esc>"),
+            "foo\n" +
+            "\n" +
+            "bar\n",
+            "fo_quux_o\n" +
+            "\n" +
+            "ba_quux_r\n");
+  }
+
+  // VIM-1379 |CTRL-V| |j| |v_b_I|
+  public void testInsertVisualBlockWithShorterLineInTheMiddle() {
+    doTest(parseKeys("ll", "<C-V>", "jjI", "_quux_", "<Esc>"),
+            "foo\n" +
+            "x\n" +
+            "bar\n",
+            "fo_quux_o\n" +
+            "x\n" +
+            "ba_quux_r\n");
+  }
+
+  // VIM-1379 |CTRL-V| |j| |v_b_c|
+  public void testChangeVisualBlockWithEmptyLineInTheMiddle() {
+    doTest(parseKeys("ll", "<C-V>", "ljjc", "_quux_", "<Esc>"),
+            "foo foo\n" +
+            "\n" +
+            "bar bar\n",
+            "fo_quux_foo\n" +
+            "\n" +
+            "ba_quux_bar\n");
+  }
+
+  // VIM-1379 |CTRL-V| |j| |v_b_c|
+  public void testChangeVisualBlockWithShorterLineInTheMiddle() {
+    doTest(parseKeys("ll", "<C-V>", "ljjc", "_quux_", "<Esc>"),
+            "foo foo\n" +
+            "x\n" +
+            "bar bar\n",
+            "fo_quux_foo\n" +
+            "x\n" +
+            "ba_quux_bar\n");
+  }
+
   // VIM-845 |CTRL-V| |x|
   public void testDeleteVisualBlockOneCharWide() {
     configureByText("foo\n" +
@@ -575,5 +629,20 @@ public class ChangeActionTest extends VimTestCase {
 
   public void testRepeatChangeWordDoesNotBreakNextRepeatFind() {
     doTest(parseKeys("fXcfYPATATA<Esc>fX.;."), "<caret>aaaaXBBBBYaaaaaaaXBBBBYaaaaaaXBBBBYaaaaaaaa\n", "aaaaPATATAaaaaaaaPATATAaaaaaaPATATAaaaaaaaa\n");
+  }
+
+  // VIM-1110 |CTRL-V| |v_b_i| |zc|
+  public void testBlockInsertAfterFolds() {
+    configureByJavaText("<caret>/**\n" +
+                        " * Something to fold.\n" +
+                        " */\n" +
+                        "foo\n" +
+                        "bar\n");
+    typeText(parseKeys("zc", "j", "<C-V>", "j", "I", "X", "<Esc>"));
+    myFixture.checkResult("/**\n" +
+                          " * Something to fold.\n" +
+                          " */\n" +
+                          "<caret>Xfoo\n" +
+                          "Xbar\n");
   }
 }